# specs/001-commit-timeline/contracts/api.yaml
openapi: 3.1.0
info:
  title: Cistory API
  description: GitHub Commit Timeline Service API
  version: 1.0.0
servers:
  - url: /api
    description: API Routes (Next.js App Router)
paths:
...
  /timeline/commits/{commitId}/summary:
    post:
      summary: AI 요약 재생성 요청
      tags: [Timeline]
      security:
        - cookieAuth: []
      parameters:
        - name: commitId
          in: path
          required: true
          schema:
            type: string
      responses:
        '202':
          description: 요약 생성 요청됨
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "요약 생성이 시작되었습니다"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          description: 재시도 횟수 초과

  # ============ Sync ============
  /sync:
    post:
      summary: 수동 동기화 시작
      tags: [Sync]
      security:
        - cookieAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                repositoryId:
                  type: string
                  description: 특정 레포만 동기화 (없으면 전체)
      responses:
        '202':
          description: 동기화 시작됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SyncJob'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: 이미 동기화 진행 중
...
  /sync/jobs:
    get:
      summary: 최근 동기화 작업 목록
      tags: [Sync]
      security:
        - cookieAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 동기화 작업 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SyncJob'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ============ Settings ============
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: session
  responses:
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: "Unauthorized"
  schemas:
...
    SyncJob:
      type: object
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, fetching, summarizing, completed, failed]
        triggerType:
          type: string
          enum: [manual, scheduled]
        totalCommits:
          type: integer
        processedCommits:
          type: integer
        errorMessage:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
...